feat: recordFormEntry
